requirements.txt
httpx>=0.27
asyncio-pool>=0.6
asyncpg>=0.29
rapidfuzz>=3.6
pandas>=2.1
python-dotenv>=1.0
beautifulsoup4>=4.12
gspread>=6.0
google-auth>=2.27
tqdm>=4.66
click>=8.1
